client: Add IoTHubClientBuilder::pnp_model_id()
The idea with PNP is that connecting devices can announce a model ID
which should be a valid DTMI when connecting. This functionality is only
really supported for MQTT.

The IoTHubClientBuilder accepts a str that it attempts to turn into a
Dtmi instance which can only be constructed if it passes validation.